+++ xknx/core/state_updater.py
@@ -209,9 +209,8 @@ class StateUpdater:
         if state == XknxConnectionState.CONNECTED:
             if not self.started:
                 self._start()
-        else:
-            if self.started:
-                self._stop()
+        elif self.started:
+            self._stop()
 
 
 class StateTrackerType(Enum):
@@ -266,7 +265,7 @@ class _StateTracker:
     async def _update_loop(self) -> None:
         """Wait for the update_interval to expire. Endless loop for updating states."""
         # for StateUpdaterType.EXPIRE:
-        #   on successfull read the while loop gets canceled when the callback calls update_received()
+        #   on successful read the while loop gets canceled when the callback calls update_received()
         #   when no telegram was received it will try again endlessly
         while True:
             await asyncio.sleep(self.update_interval)